src/domain/@shared/events/event-dispatcher.spec.ts
import SendEmailWhenProductIsCreatedhandler from "../../product/event/handler/send-email-when-product-created-handler";
import { EventDispatcher } from "./event-dispatcher";

describe("Domain Events Tests", () => {
  it("Should register an event handler", () => {
    const eventDispatcher = new EventDispatcher();
    const eventHandler = new SendEmailWhenProductIsCreatedhandler();

    eventDispatcher.register("ProductCreatedEvent", eventHandler);

    expect(
      eventDispatcher.getEventHandlers["ProductCreatedEvent"]
    ).toBeDefined();

    expect(eventDispatcher.getEventHandlers["ProductCreatedEvent"][0]).toBe(1);

    expect(
      eventDispatcher.getEventHandlers["ProductCreatedEvent"][0]
    ).toMatchObject(eventHandler);
  });
});
